refactor: Reorganize files according to data/domain/UI layer
