+++ app/src/main/java/campus/tech/kakao/map/data/PlaceDataModel.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data
 
 data class PlaceDataModel(
     val name: String,
