+++ app/src/main/java/campus/tech/kakao/map/data/PlaceDatabaseAccess.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data
 
 import android.content.ContentValues
 import android.content.Context
