+++ app/src/main/java/campus/tech/kakao/map/data/PlaceDatabaseHelper.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
