+++ app/src/main/java/campus/tech/kakao/map/data/PlaceRepository.kt
@@ -1,13 +1,17 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data
 
 import android.util.Log
+import campus.tech.kakao.map.BuildConfig
+import campus.tech.kakao.map.domain.RetrofitInstance
+import campus.tech.kakao.map.domain.RetrofitLocalKeywordService
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
 class PlaceRepository {
 
-    private val retrofitLocalKeywordService: RetrofitLocalKeywordService = RetrofitInstance.retrofitLocalKeywordService
+    private val retrofitLocalKeywordService: RetrofitLocalKeywordService =
+        RetrofitInstance.retrofitLocalKeywordService
 
     fun searchPlace(
         keyword: String,
