+++ app/src/main/java/campus/tech/kakao/map/domain/RetrofitInstance.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.domain
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -11,6 +11,8 @@ object RetrofitInstance {
         .addConverterFactory(GsonConverterFactory.create())
         .build()
 
-    val retrofitLocalCategoryService: RetrofitLocalCategoryService = retrofit.create(RetrofitLocalCategoryService::class.java)
-    val retrofitLocalKeywordService: RetrofitLocalKeywordService = retrofit.create(RetrofitLocalKeywordService::class.java)
+    val retrofitLocalCategoryService: RetrofitLocalCategoryService = retrofit.create(
+        RetrofitLocalCategoryService::class.java)
+    val retrofitLocalKeywordService: RetrofitLocalKeywordService = retrofit.create(
+        RetrofitLocalKeywordService::class.java)
 }
