+++ app/src/main/java/campus/tech/kakao/map/presentation/KaKaoMapApplication.kt
@@ -1,6 +1,7 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.presentation
 
 import android.app.Application
+import campus.tech.kakao.map.BuildConfig
 import com.kakao.vectormap.KakaoMapSdk
 
 class KaKaoMapApplication : Application() {
