+++ app/src/main/java/campus/tech/kakao/map/ui/ModalBottomSheet.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
