+++ app/src/main/java/campus/tech/kakao/map/ui/PlaceRecyclerViewAdapter.kt
@@ -1,10 +1,12 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.ui
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import campus.tech.kakao.map.data.PlaceDataModel
+import campus.tech.kakao.map.R
 
 class PlaceRecyclerViewAdapter(
     private val places: MutableList<PlaceDataModel>,
