+++ app/src/main/java/campus/tech/kakao/map/ui/SearchRecyclerViewAdapter.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.ui
 
 import android.view.LayoutInflater
 import android.view.View
@@ -8,6 +8,8 @@ import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import campus.tech.kakao.map.data.PlaceDataModel
+import campus.tech.kakao.map.R
 
 class SearchRecyclerViewAdapter(
     private val places: MutableList<PlaceDataModel>,
