+++ app/src/main/res/layout/search_layout.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlaceActivity">
+    tools:context=".presentation.PlaceActivity">
 
     <LinearLayout
         android:id="@+id/llSearch"
